requirements.txt
torch
torchvision
diffusers
transformers
accelerate
pillow
opencv-python
fastapi
uvicorn
huggingface-hub


